requirements.txt
python-dotenv==1.0.1
langchain-text-splitters
langchain-community
langgraph
langchain[google-genai]
langchain_google_genai
langchain-qdrant
bs4
streamlit
pypdf
dependency-injector
mediatr